Remove unused entities Refactor JPA models
