+++ .gitignore
@@ -33,3 +33,4 @@ build/
 .vscode/
 src/main/resources/application.properties
 .jpb
+.run
