--- src/main/java/com/andrewjoel/blogic/repository/MetricRepository.java
@@ -1,9 +1,0 @@
-package com.andrewjoel.blogic.repository;
-
-import com.andrewjoel.blogic.models.jpa.Metric;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MetricRepository extends JpaRepository<Metric, Long> {
-}
