--- src/main/java/com/andrewjoel/blogic/repository/PostRepository.java
@@ -1,9 +1,0 @@
-package com.andrewjoel.blogic.repository;
-
-import com.andrewjoel.blogic.models.jpa.Post;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PostRepository extends JpaRepository<Post, Long> {
-}
