+++ src/main/java/org/blogic/BlogicApplication.java
@@ -1,4 +1,4 @@
-package com.andrewjoel.blogic;
+package org.blogic;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class BlogicApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(BlogicApplication.class, args);
-	}
+    public static void main(final String[] args) {
+        SpringApplication.run(BlogicApplication.class, args);
+    }
 
 }
