+++ src/main/java/org/blogic/core/models/enums/Status.java
@@ -1,4 +1,4 @@
-package com.andrewjoel.blogic.models.enums;
+package org.blogic.core.models.enums;
 
 public enum Status {
     DRAFT,
