+++ src/main/java/org/blogic/core/models/enums/UserType.java
@@ -1,0 +1,6 @@
+package org.blogic.core.models.enums;
+
+public enum UserType {
+    APP_USER,
+    SERVICE_USER
+}
