+++ src/main/java/org/blogic/core/models/jpa/AbstractBaseEntity.java
@@ -1,0 +1,30 @@
+package org.blogic.core.models.jpa;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+
+import javax.persistence.*;
+import java.io.Serial;
+import java.io.Serializable;
+import java.time.LocalDateTime;
+
+@MappedSuperclass
+@Getter
+@Setter
+public abstract class AbstractBaseEntity implements Serializable {
+    @Serial
+    private static final long serialVersionUID = -2313432625968445331L;
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(updatable = false, nullable = false)
+    private Long id;
+
+    @CreationTimestamp
+    private LocalDateTime createdAt;
+
+    @UpdateTimestamp
+    private LocalDateTime modifiedAt;
+}
