+++ src/main/java/org/blogic/core/repository/AuthRepository.java
@@ -1,0 +1,9 @@
+package org.blogic.core.repository;
+
+import org.blogic.core.models.jpa.Auth;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AuthRepository extends JpaRepository<Auth, Long> {
+}
